legacy/requirements.txt
asgiref==3.7.2
astroid==3.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.7
Django==4.2.7
django-rosetta==0.9.9
gunicorn==21.2.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
packaging==23.2
platformdirs==4.0.0
polib==1.2.0
pylint==3.0.2
requests==2.31.0
RPi.GPIO==0.7.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
